@@ -10,40 +10,53 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@hookform/resolvers": "^5.2.2",
     "@radix-ui/react-accordion": "^1.2.1",
-    "@radix-ui/react-separator": "^1.1.0",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-toast": "^1.2.2",
-    "@react-three/fiber": "^8.17.10",
+    "@radix-ui/react-avatar": "^1.1.10",
+    "@radix-ui/react-checkbox": "^1.3.3",
+    "@radix-ui/react-label": "^2.1.7",
+    "@radix-ui/react-navigation-menu": "^1.2.14",
+    "@radix-ui/react-separator": "^1.1.7",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@radix-ui/react-switch": "^1.2.6",
+    "@radix-ui/react-toast": "^1.2.15",
+    "@radix-ui/react-tooltip": "^1.2.8",
     "@tailwindcss/typography": "^0.5.15",
     "bootstrap": "^5.3.3",
     "bufferutil": "^4.0.9",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "flowbite": "3.1.2",
+    "flowbite-react": "0.12.9",
     "framer-motion": "^11.11.9",
+    "input-otp": "^1.4.2",
     "jquery": "^3.7.1",
     "lucide-react": "^0.453.0",
     "mini-svg-data-uri": "^1.4.4",
-    "next": "^14.2.25",
+    "motion": "^12.23.22",
+    "next": "^15",
     "next-themes": "^0.3.0",
     "pocketbase": "^0.21.5",
-    "react": "^18",
+    "react": "18.3.1",
     "react-bootstrap": "^2.10.5",
-    "react-dom": "^18",
+    "react-dom": "18.3.1",
+    "react-hook-form": "^7.63.0",
     "react-icons": "^5.3.0",
     "tailwind-merge": "^2.5.4",
-    "three": "^0.169.0",
     "use-debounce": "^10.0.4",
     "utf-8-validate": "^6.0.5",
-    "ws": "^8.18.3"
+    "ws": "^8.18.3",
+    "zod": "^4.1.11"
   },
   "devDependencies": {
+    "@tailwindcss/postcss": "^4.0.0",
     "@types/node": "24.1.0",
     "@types/three": "^0.169.0",
     "eslint": "^8",
-    "eslint-config-next": "14.2.15",
+    "eslint-config-next": "^15",
     "postcss": "^8",
-    "tailwindcss": "^3.4.1"
+    "shadcn": "^3.3.1",
+    "tailwindcss": "^4.0.0"
   },
   "description": "This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).",
   "main": "tailwind.config.js",
