@@ -1,29 +1,34 @@
 'use client';
 
-
-import { Boxes } from "./ui/background-boxes";
+import { HeroHighlight, Highlight } from "@/components/ui/hero-highlight";
+import { Button } from "@/components/ui/button";
 
 export default function Hero() {
+  const badges = ["Linux-first", "Next.js", "Node.js", "Networking", "Automation"];
+
   return (
-    <div className="relative w-full h-full overflow-hidden bg-frappe-crust flex flex-col items-center justify-center">
-      <div className="absolute inset-0 w-full h-full bg-frappe-crust z-10 [mask-image:radial-gradient(transparent,white)] pointer-events-none" />
-      
-      <Boxes />
-      <div className="text-center max-w-4xl px-4 relative z-20">
-        <h1 className="text-4xl md:text-6xl lg:text-7xl font-bold text-white leading-tight lg:leading-tight mb-4">
-          whoami
-        </h1>
-        <h2 className="text-3xl md:text-5xl lg:text-6xl font-bold mb-10 flex flex-col items-center">
-          <span className="text-white bg-frappe-surface0 inline-block mb-2 px-4 py-2 rounded-lg">
-          /home/jason
-          </span>
-        </h2>
-        <a 
-          className="px-8 py-4 text-lg rounded-full bg-frappe-blue text-white hover:bg-frappe-sapphire transition-colors duration-300"
-          href="#Featured"
-        >
-          Let&apos;s begin
-        </a>
+    <div className="relative w-full h-full flex items-center">
+      <div className="w-full px-6">
+        <div className="mx-auto max-w-5xl text-center">
+          <HeroHighlight containerClassName="rounded-xl">
+            <div>
+              <h1 className="text-3xl md:text-5xl font-bold tracking-tight">
+                <Highlight>whoisjason</Highlight>
+              </h1>
+              <p className="mt-3 text-sm text-muted-foreground">
+                Minimal, reliable web apps. UNIX discipline. Full‑stack JavaScript.
+              </p>
+              <div className="mt-6 flex items-center justify-center gap-3">
+                <Button asChild>
+                  <a href="#featured">Featured</a>
+                </Button>
+                <Button variant="outline" asChild>
+                  <a href="#projects">Projects</a>
+                </Button>
+              </div>
+            </div>
+          </HeroHighlight>
+        </div>
       </div>
     </div>
   );
