@@ -37,10 +37,10 @@ const toastVariants = cva(
 
 const Toast = React.forwardRef(({ className, variant, ...props }, ref) => {
   return (
-    (<ToastPrimitives.Root
+    <ToastPrimitives.Root
       ref={ref}
       className={cn(toastVariants({ variant }), className)}
-      {...props} />)
+      {...props} />
   );
 })
 Toast.displayName = ToastPrimitives.Root.displayName
